refactor(viz): crystallize plotting templates — split monolith into per-section files
Phase 2 of the multi-phase cleanup. Carves the 1180-line
pipeline_plots.py + 428-line lrg_multiscale_plots.py + the two
sankey backends into per-concept section files behind a central
style module. Library output is unchanged; structure and
discoverability change. ruff clean, pytest 119/119, one notebook
executed end-to-end with zero errors.

New foundation files
--------------------

- multifunbrain/visualization/style.py
  PALETTES (material/spectral/signed/contrast), FIGSIZE defaults,
  rc_context(profile) for paper/slide rcParams. Single source of
  truth for the project's plot colours and sizes.

- multifunbrain/visualization/plotlib/base.py
  ensure_axes(ax, figsize=, **subplots_kwargs) — the standard
  if-None-create-else-get_figure guard. apply_decorations(ax,
  title=, xlabel=, ylabel=, legend=, grid=) for the kwargs every
  grid-composable plot accepts.

- multifunbrain/visualization/plotlib/_helpers.py
  Package-private helpers used across section files: _get_section,
  _resolve_filter, _signed_laplacian_embedding, _network_layout,
  _nearest_tau_index.

Per-section plot files (split from pipeline_plots / lrg_multiscale)
------------------------------------------------------------------

- descriptive.py: plot_weight_distribution, plot_eigenvalue_spectrum,
  plot_signed_laplacian_spectrum, plot_signed_balance,
  plot_correlation_matrix.
- filtering.py: plot_percolation_curve, plot_filtered_comparison.
- network.py: plot_node_metrics, plot_signed_network, plot_network.
- lrg.py: plot_lrg_entropy / plot_lrg_dendrogram / plot_lrg_psi /
  plot_lrg_partition_network / plot_lrg_sankey (from pipeline_plots)
  + all 8 from lrg_multiscale_plots (specific_heat, dendrogram_with_psi,
  psi_curve, rmi_curve, partition_flow, tanglegram, metastable_overlay,
  specific_heat_overlay).
- grids.py: plot_results_grid (canonical multi-panel composer),
  plot_pipeline_summary (six-panel single-result overview).
- sankey.py: plot_sankey(partitions, taus, backend="plotly"|"matplotlib")
  merging both backends behind one entry point;
  plot_sankey_matplotlib preserved as a back-compat alias.

Switch + deletions
------------------

- plotlib/__init__.py rewritten to import from the new layout.
  Public __all__ preserved exactly so callers (CLI, notebooks,
  scripts, the `from multifunbrain.notebook import *` ergonomic
  namespace) keep working.
- Deleted: pipeline_plots.py (1180 L), lrg_multiscale_plots.py
  (428 L), sankey_matplotlib.py (137 L), sankey_plotly.py (63 L).

Crystallized templates (initial set)
------------------------------------

Three representative plots refactored to pull colours/sizes from
the central style module instead of inline hex strings, proving the
pattern future plots will follow:

- CORRELATION_HEATMAP (descriptive.plot_correlation_matrix):
  FIGSIZE["square"] for the default canvas.
- DENDROGRAM (lrg.plot_lrg_dendrogram): threshold-line colour
  from PALETTES["material"]["danger_light"]; wide figsize.
- SPECIFIC_HEAT (lrg.plot_specific_heat): trace, τ′ and τ* lines
  from PALETTES["material"] primary_dark / accent / success.

The remaining ~25 plot helpers keep their original hardcoded
colours for now; follow-up PRs can convert them as the user
reviews each pattern.
