+++ multifunbrain/visualization/plotlib/__init__.py
@@ -1,8 +1,35 @@
 """Plotting utilities for :mod:`multifunbrain`.
 
-This package centralises heavy plotting imports so that submodules can
-reuse shared handles (for example ``plt`` or ``go``) without repeating
-the import cost in every notebook run.
+This package centralises the heavy matplotlib / plotly imports so that
+section files can reuse shared handles (``plt``, ``go``, ``Polygon``,
+``Rectangle``) without paying the import cost in each notebook cell.
+
+Section files (one per pipeline concept) hold the actual plot functions:
+
+- :mod:`~multifunbrain.visualization.plotlib.descriptive` — weight
+  distribution, eigenvalue spectrum, signed Laplacian, correlation
+  matrix heatmap.
+- :mod:`~multifunbrain.visualization.plotlib.filtering` — percolation
+  curves, filtered network comparison.
+- :mod:`~multifunbrain.visualization.plotlib.network` — node metrics,
+  signed/unsigned network drawings.
+- :mod:`~multifunbrain.visualization.plotlib.lrg` — entropy/specific
+  heat, dendrograms, PSI/RMI, partition flow, tanglegram, metastable
+  overlay.
+- :mod:`~multifunbrain.visualization.plotlib.grids` — ``plot_results_grid``
+  and ``plot_pipeline_summary``.
+- :mod:`~multifunbrain.visualization.plotlib.sankey` — Sankey diagrams
+  (plotly + matplotlib backends behind one entry point).
+- :mod:`~multifunbrain.visualization.plotlib.entropy` — dual-axis
+  entropy + C atomic helper.
+- :mod:`~multifunbrain.visualization.plotlib.colorbars` — colorbar
+  layout utilities.
+- :mod:`~multifunbrain.visualization.plotlib.base` — ``ensure_axes``
+  and ``apply_decorations`` boilerplate helpers used by all plots.
+
+Style defaults (palettes, figsizes, rcParams profiles) live one level
+up in :mod:`multifunbrain.visualization.style` — pull colour and size
+constants from there rather than hardcoding hex strings.
 """
 
 from __future__ import annotations
@@ -19,10 +46,25 @@ from matplotlib.colors import (
 from matplotlib.patches import Polygon, Rectangle
 from mpl_toolkits.mplot3d import Axes3D
 
+from .base import apply_decorations, ensure_axes
 from .colorbars import imshow_colorbar_caxdivider
+from .descriptive import (
+    plot_correlation_matrix,
+    plot_eigenvalue_spectrum,
+    plot_signed_balance,
+    plot_signed_laplacian_spectrum,
+    plot_weight_distribution,
+)
 from .entropy import plot_entropy_and_C
-from .lrg_multiscale_plots import (
+from .filtering import plot_filtered_comparison, plot_percolation_curve
+from .grids import plot_pipeline_summary, plot_results_grid
+from .lrg import (
     plot_dendrogram_with_psi,
+    plot_lrg_dendrogram,
+    plot_lrg_entropy,
+    plot_lrg_partition_network,
+    plot_lrg_psi,
+    plot_lrg_sankey,
     plot_metastable_overlay,
     plot_partition_flow,
     plot_psi_curve,
@@ -31,65 +73,56 @@ from .lrg_multiscale_plots import (
     plot_specific_heat_overlay,
     plot_tanglegram,
 )
-from .pipeline_plots import (
-    plot_correlation_matrix,
-    plot_eigenvalue_spectrum,
-    plot_filtered_comparison,
-    plot_lrg_dendrogram,
-    plot_lrg_entropy,
-    plot_lrg_partition_network,
-    plot_lrg_psi,
-    plot_lrg_sankey,
-    plot_network,
-    plot_node_metrics,
-    plot_percolation_curve,
-    plot_pipeline_summary,
-    plot_results_grid,
-    plot_signed_balance,
-    plot_signed_laplacian_spectrum,
-    plot_signed_network,
-    plot_weight_distribution,
-)
-from .sankey_matplotlib import plot_sankey_matplotlib
-from .sankey_plotly import plot_sankey
+from .network import plot_network, plot_node_metrics, plot_signed_network
+from .sankey import plot_sankey, plot_sankey_matplotlib
 
 __all__ = [
+    # Public matplotlib / plotly handles (re-exported for notebook convenience)
+    "Axes3D",
+    "BoundaryNorm",
+    "LinearSegmentedColormap",
+    "ListedColormap",
+    "Polygon",
+    "Rectangle",
+    "TwoSlopeNorm",
+    "apply_decorations",
+    "ensure_axes",
+    "go",
+    "gridspec",
+    "imshow_colorbar_caxdivider",
+    "plt",
+    # descriptive
     "plot_correlation_matrix",
-    "plot_dendrogram_with_psi",
     "plot_eigenvalue_spectrum",
     "plot_entropy_and_C",
+    "plot_signed_balance",
+    "plot_signed_laplacian_spectrum",
+    "plot_weight_distribution",
+    # filtering
     "plot_filtered_comparison",
+    "plot_percolation_curve",
+    # network
+    "plot_network",
+    "plot_node_metrics",
+    "plot_signed_network",
+    # lrg (per-PipelineResult + multiscale)
+    "plot_dendrogram_with_psi",
     "plot_lrg_dendrogram",
     "plot_lrg_entropy",
     "plot_lrg_partition_network",
     "plot_lrg_psi",
     "plot_lrg_sankey",
     "plot_metastable_overlay",
-    "plot_network",
-    "plot_node_metrics",
     "plot_partition_flow",
-    "plot_percolation_curve",
-    "plot_pipeline_summary",
     "plot_psi_curve",
-    "plot_results_grid",
     "plot_rmi_curve",
-    "plot_sankey",
-    "plot_sankey_matplotlib",
-    "plot_signed_balance",
-    "plot_signed_laplacian_spectrum",
-    "plot_signed_network",
     "plot_specific_heat",
     "plot_specific_heat_overlay",
     "plot_tanglegram",
-    "plot_weight_distribution",
-    "plt",
-    "go",
-    "Rectangle",
-    "Polygon",
-    "Axes3D",
-    "BoundaryNorm",
-    "LinearSegmentedColormap",
-    "ListedColormap",
-    "TwoSlopeNorm",
-    "imshow_colorbar_caxdivider",
+    # grids
+    "plot_pipeline_summary",
+    "plot_results_grid",
+    # sankey
+    "plot_sankey",
+    "plot_sankey_matplotlib",
 ]
